Refactor development workflow for flask app
